fix(broker): some doubles forgotten

+++ broker/unified_sql/test/perfdata.cc
@@ -196,9 +196,7 @@ TEST(UnifiedSqlPerfdata, DefaultCtor) {
 
 class UnifiedSqlParserParsePerfdata : public testing::Test {
  public:
-  void SetUp() override {
-    config::applier::init(0, "test_broker", 0);
-  }
+  void SetUp() override { config::applier::init(0, "test_broker", 0); }
   void TearDown() override { config::applier::deinit(); };
 };
 
@@ -237,9 +235,9 @@ TEST_F(UnifiedSqlParserParsePerfdata, Simple2) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(18.0);
   expected.unit("%");
-  expected.warning(std::numeric_limits<double>::infinity());
+  expected.warning(std::numeric_limits<float>::infinity());
   expected.warning_low(15.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(10.0);
   expected.min(0.0);
   expected.max(100.0);
@@ -263,7 +261,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(2.45698);
   expected.unit("s");
-  expected.max(std::numeric_limits<double>::infinity());
+  expected.max(std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ++it;
 
@@ -275,7 +273,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.unit("B/s");
   expected.warning(5.0);
   expected.warning_low(0.0);
-  expected.min(-std::numeric_limits<double>::infinity());
+  expected.min(-std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ++it;
 
@@ -295,7 +293,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value(1234.0);
   expected.warning(10.0);
   expected.warning_low(0.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(11.0);
   ASSERT_TRUE(expected == *it);
   ++it;
@@ -306,7 +304,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::counter);
   expected.value(1234.0);
   expected.warning(10.0);
-  expected.warning_low(-std::numeric_limits<double>::infinity());
+  expected.warning_low(-std::numeric_limits<float>::infinity());
   expected.critical(30.0);
   expected.critical_low(20.0);
   ASSERT_TRUE(expected == *it);
@@ -329,10 +327,10 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(9.0);
   expected.unit("queries_per_second");
-  expected.warning(std::numeric_limits<double>::infinity());
+  expected.warning(std::numeric_limits<float>::infinity());
   expected.warning_low(10.0);
   expected.warning_mode(true);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(5.0);
   expected.critical_mode(true);
   expected.min(0.0);
@@ -444,7 +442,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(2.45698);
   expected.unit("s");
-  expected.max(std::numeric_limits<double>::infinity());
+  expected.max(std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
   ++it;
@@ -457,7 +455,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.unit("B/s");
   expected.warning(5.0);
   expected.warning_low(0.0);
-  expected.min(-std::numeric_limits<double>::infinity());
+  expected.min(-std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
   ++it;
@@ -489,7 +487,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.value(1234.17);
   expected.warning(10.0);
   expected.warning_low(0.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(11.0);
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
@@ -501,7 +499,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.value_type(misc::perfdata::counter);
   expected.value(1234.147);
   expected.warning(10.0);
-  expected.warning_low(-std::numeric_limits<double>::infinity());
+  expected.warning_low(-std::numeric_limits<float>::infinity());
   expected.critical(30.0);
   expected.critical_low(20.0);
   ASSERT_TRUE(expected == *it);
@@ -543,7 +541,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, BadMetric) {
   int i = 1;
   for (auto& p : lst) {
     ASSERT_EQ(p.name(), fmt::format("user{}", i));
-    ASSERT_EQ(p.value(), static_cast<double>(i));
+    ASSERT_EQ(p.value(), static_cast<float>(i));
     ++i;
   }
 }
@@ -556,7 +554,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, BadMetric1) {
   int i = 1;
   for (auto& p : lst) {
     ASSERT_EQ(p.name(), fmt::format("user{}", i));
-    ASSERT_EQ(p.value(), static_cast<double>(i));
+    ASSERT_EQ(p.value(), static_cast<float>(i));
     ++i;
   }
 }
